+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed a bug where some default callbacks could be added twice if given a different name by the user.
+- Fixed a bug where some `Trainer` bookkeeping tasks may not complete before `.fit()` returns.
 
 ## [v1.6.1](https://github.com/allenai/OLMo-core/releases/tag/v1.6.1) - 2024-11-06